#5: Further fixes to repr(), ChainedUnit
